+++ test-projects/login/CHANGELOG.md
@@ -1,3 +1,5 @@
+# @keystone-alpha/cypress-project-login
+
 # @voussoir/cypress-project-login
 
 ## 2.0.0